+++ OtpishAI_Backend_Codes/src/main/java/otpishAI/otpishAI_Backend/repository/ProductRepository.java
@@ -8,7 +8,7 @@ import otpishAI.otpishAI_Backend.entity.Product;
 public interface ProductRepository extends JpaRepository<Product, String> {
 
     Page<Product> findAllByOrderByPaymentDesc(Pageable pageable);
-    Product findByProductNum(Integer productNum);
+    Product findByProductNum(Long productNum);
 
 
 }
